# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AgilePlus",
    "description": "A web application for intuitive and simple project management. It helps to plan, track and manages a team's workflow effectively.",
    "version": "1.0.0",
    "contact": {
      "email": "skarthikayan7@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "requestBodies": {
...
      "backlogDetails": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/backlog"
            }
          }
        }
      },
...
      "sprintDetails": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/sprint"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "backlogList": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "success"
                },
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/backlog"
                      },
                      {
                        "$ref": "#/components/schemas/default"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "backlogDetails": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "success"
                },
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/backlog"
                    },
                    {
                      "$ref": "#/components/schemas/default"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "notFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "failed"
                },
                "success": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        }
      },
      "internalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "failed"
                },
                "success": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        }
      },
...
      "sprintList": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "success"
                },
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/sprint"
                      },
                      {
                        "$ref": "#/components/schemas/default"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "sprintDetails": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "success"
                },
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/sprint"
                    },
                    {
                      "$ref": "#/components/schemas/default"
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "backlog": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Backlog Name"
          },
          "story_points": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "string",
            "example": "NEW",
            "enum": [
              "NEW",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "sprint_id": {
            "type": "number",
            "example": 1
          },
          "user_id": {
            "type": "number",
            "example": 1
          }
        }
      },
      "default": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "created_at": {
            "type": "string",
            "example": "2024-08-21T19:13:08.440Z"
          },
          "updated_at": {
            "type": "string",
            "example": "2024-08-21T19:13:08.440Z"
          }
        }
      },
...
      "sprint": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sprint name"
          },
          "project_id": {
            "type": "number",
            "example": 1
          },
          "start_date": {
            "type": "string",
            "example": "2024-08-21T19:13:08.440Z"
          },
          "end_date": {
            "type": "string",
            "example": "2024-08-21T19:13:08.440Z"
          },
          "status": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/backlogs": {
      "get": {
        "tags": [
          "Backlog"
        ],
        "summary": "List all Backlogs",
        "responses": {
          "200": {
            "$ref": "#/components/responses/backlogList"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Backlog"
        ],
        "summary": "Create new backlog",
        "responses": {
          "200": {
            "$ref": "#/components/responses/backlogDetails"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        },
        "requestBody": {
          "description": "Create a new backlog",
          "$ref": "#/components/requestBodies/backlogDetails",
          "required": true
        }
      }
    },
    "/backlogs/{id}": {
      "get": {
        "tags": [
          "Backlog"
        ],
        "summary": "Show backlog Details",
        "responses": {
          "200": {
            "$ref": "#/components/responses/backlogDetails"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of backlog to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ]
      },
      "patch": {
        "tags": [
          "Backlog"
        ],
        "summary": "Edit Backlog Details",
        "responses": {
          "200": {
            "$ref": "#/components/responses/backlogDetails"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of backlog to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Create a new backlog",
          "$ref": "#/components/requestBodies/backlogDetails",
          "required": true
        }
      },
      "delete": {
        "tags": [
          "Backlog"
        ],
        "summary": "Delete Backlog Details",
        "responses": {
          "200": {
            "$ref": "#/components/responses/backlogDetails"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of backlog to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ]
      }
    },
...
    "/sprints": {
      "get": {
        "tags": [
          "Sprint"
        ],
        "summary": "List all Sprints",
        "responses": {
          "200": {
            "$ref": "#/components/responses/sprintList"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Sprint"
        ],
        "summary": "Create new sprint",
        "responses": {
          "200": {
            "$ref": "#/components/responses/sprintDetails"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        },
        "requestBody": {
          "description": "Create a new sprint",
          "$ref": "#/components/requestBodies/sprintDetails",
          "required": true
        }
      }
    },
    "/sprints/{id}": {
      "get": {
        "tags": [
          "Sprint"
        ],
        "summary": "Show sprint Details",
        "responses": {
          "200": {
            "$ref": "#/components/responses/sprintDetails"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of sprint to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ]
      },
      "patch": {
        "tags": [
          "Sprint"
        ],
        "summary": "Edit Sprint Details",
        "responses": {
          "200": {
            "$ref": "#/components/responses/sprintDetails"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of sprint to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Create a new sprint",
          "$ref": "#/components/requestBodies/sprintDetails",
          "required": true
        }
      },
      "delete": {
        "tags": [
          "Sprint"
        ],
        "summary": "Delete Sprint Details",
        "responses": {
          "200": {
            "$ref": "#/components/responses/sprintDetails"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "500": {
            "$ref": "#/components/responses/internalServerError"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of sprint to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ]
      }
    },
...
  },
...
}